requirements.txt
PyQt5
fuzzywuzzy
num_to_rus
requests
telethon
playsound
sounddevice
grpcio-tools
PyAudio
aiohttp
pvporcupine
python-Levenshtein
wikipedia
python-dotenv
python-cbr
vosk
pydub
pyttsx3
shazamio
SpeechRecognition